bfcp: fix sub-expression overflow warning

+++ src/bfcp/attr.c
@@ -130,7 +130,7 @@ static int attr_encode(struct mbuf *mb, bool mand, enum bfcp_attrib type,
 	mb->pos = start;
 	err |= mbuf_write_u8(mb, (type<<1) | (mand ? 1 : 0));
 	err |= mbuf_write_u8(mb, len);
-	mb->pos += (len - BFCP_ATTR_HDR_SIZE);
+	mb->pos += (size_t)(len - BFCP_ATTR_HDR_SIZE);
 
 	/* padding */
 	while ((mb->pos - start) & 0x03)
